Fix budget model attributes
