+++ src/app/layout/budget/budget.component.ts
@@ -13,7 +13,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { BudgetService } from '../../services/budget.service';
 import { CreateBudgetDto } from '../../services/dtos/event/create-budget.dto';
-import { Budget } from '../../model/budget';
+import { Budget } from '../../model/budget/budget';
 import { DeleteDialogComponent } from '../../dialog/delete-dialog/delete-dialog.component';
 import { MatDialog } from '@angular/material/dialog';
 import { ServiceProductService } from '../../services/service-product/service-product.service';
