+++ src/app/model/budget/booking.ts
@@ -1,0 +1,9 @@
+import { Service } from "../service-product/service";
+
+export interface Booking {
+    id: number;
+    service: Service;
+    price: number;
+    date: Date;
+    duration: number;
+}
