+++ src/app/model/budget/purchase.ts
@@ -1,0 +1,7 @@
+import { Product } from "../service-product/product";
+
+export interface Purchase {
+    id: number;
+    product: Product;
+    price: number;
+}
